move varverletlists to separate directory
