+++ src/autopas/containers/verletListsCellBased/VarVerletLists/neighborLists/asBuild/ColorChangeObserver.h
